+++ src/io.rs
@@ -6,7 +6,7 @@
 use crate::bitmap::BitmapSlice;
 use crate::volatile_memory::copy_slice_impl::{copy_from_volatile_slice, copy_to_volatile_slice};
 use crate::{VolatileMemoryError, VolatileSlice};
-use std::io::ErrorKind;
+use std::io::{Cursor, ErrorKind, Stdout};
 use std::os::fd::AsRawFd;
 
 /// A version of the standard library's [`Read`] trait that operates on volatile memory instead of
@@ -132,6 +132,15 @@ macro_rules! impl_read_write_volatile_for_raw_fd {
             }
         }
     };
+}
+
+impl WriteVolatile for Stdout {
+    fn write_volatile<B: BitmapSlice>(
+        &mut self,
+        buf: &VolatileSlice<B>,
+    ) -> Result<usize, VolatileMemoryError> {
+        write_volatile_raw_fd(self, buf)
+    }
 }
 
 impl_read_write_volatile_for_raw_fd!(std::fs::File);
@@ -271,11 +280,87 @@ impl ReadVolatile for &[u8] {
     }
 }
 
+// WriteVolatile implementation for Vec<u8> is based upon the Write impl for Vec, which
+// defers to Vec::append_elements, after which the below functionality is modelled.
+impl WriteVolatile for Vec<u8> {
+    fn write_volatile<B: BitmapSlice>(
+        &mut self,
+        buf: &VolatileSlice<B>,
+    ) -> Result<usize, VolatileMemoryError> {
+        let count = buf.len();
+        self.reserve(count);
+        let len = self.len();
+
+        // SAFETY: Calling Vec::reserve() above guarantees the the backing storage of the Vec has
+        // length at least `len + count`. This means that self.as_mut_ptr().add(len) remains within
+        // the same allocated object, the offset does not exceed isize (as otherwise reserve would
+        // have panicked), and does not rely on address space wrapping around.
+        // In particular, the entire `count` bytes after `self.as_mut_ptr().add(count)` is
+        // contiguously allocated and valid for writes.
+        // Lastly, `copy_to_volatile_slice` correctly initialized `copied_len` additional bytes
+        // in the Vec's backing storage, and we assert this to be equal to `count`. Additionally,
+        // `len + count` is at most the reserved capacity of the vector. Thus the call to `set_len`
+        // is safe.
+        unsafe {
+            let copied_len = copy_from_volatile_slice(self.as_mut_ptr().add(len), buf, count);
+
+            assert_eq!(copied_len, count);
+            self.set_len(len + count);
+        }
+        Ok(count)
+    }
+}
+
+// ReadVolatile and WriteVolatile implementations for Cursor<T> is modelled after the standard
+// library's implementation (modulo having to inline `Cursor::remaining_slice`, as that's nightly only)
+impl<T> ReadVolatile for Cursor<T>
+where
+    T: AsRef<[u8]>,
+{
+    fn read_volatile<B: BitmapSlice>(
+        &mut self,
+        buf: &mut VolatileSlice<B>,
+    ) -> Result<usize, VolatileMemoryError> {
+        let inner = self.get_ref().as_ref();
+        let len = self.position().min(inner.len() as u64);
+        let n = ReadVolatile::read_volatile(&mut &inner[(len as usize)..], buf)?;
+        self.set_position(self.position() + n as u64);
+        Ok(n)
+    }
+
+    fn read_exact_volatile<B: BitmapSlice>(
+        &mut self,
+        buf: &mut VolatileSlice<B>,
+    ) -> Result<(), VolatileMemoryError> {
+        let inner = self.get_ref().as_ref();
+        let n = buf.len();
+        let len = self.position().min(inner.len() as u64);
+        ReadVolatile::read_exact_volatile(&mut &inner[(len as usize)..], buf)?;
+        self.set_position(self.position() + n as u64);
+        Ok(())
+    }
+}
+
+impl WriteVolatile for Cursor<&mut [u8]> {
+    fn write_volatile<B: BitmapSlice>(
+        &mut self,
+        buf: &VolatileSlice<B>,
+    ) -> Result<usize, VolatileMemoryError> {
+        let pos = self.position().min(self.get_ref().len() as u64);
+        let n = WriteVolatile::write_volatile(&mut &mut self.get_mut()[(pos as usize)..], buf)?;
+        self.set_position(self.position() + n as u64);
+        Ok(n)
+    }
+
+    // no write_all provided in standard library, since our default for write_all is based on the
+    // standard library's write_all, omitting it here as well will correctly mimic stdlib behavior.
+}
+
 #[cfg(test)]
 mod tests {
     use crate::io::{ReadVolatile, WriteVolatile};
     use crate::{VolatileMemoryError, VolatileSlice};
-    use std::io::{ErrorKind, Read, Seek, Write};
+    use std::io::{Cursor, ErrorKind, Read, Seek, Write};
     use vmm_sys_util::tempfile::TempFile;
 
     // ---- Test ReadVolatile for &[u8] ----
@@ -440,5 +525,86 @@ mod tests {
             write_4_bytes_to_5_byte_vec(input.clone(), output);
             write_5_bytes_to_file(input);
         }
+    }
+
+    #[test]
+    fn test_read_volatile_for_cursor() {
+        let read_buffer = [1, 2, 3, 4, 5, 6, 7];
+        let mut output = vec![0u8; 5];
+
+        let mut cursor = Cursor::new(read_buffer);
+
+        // Read 4 bytes from cursor to volatile slice (amount read limited by volatile slice length)
+        assert_eq!(
+            cursor
+                .read_volatile(&mut VolatileSlice::from(&mut output[..4]))
+                .unwrap(),
+            4
+        );
+        assert_eq!(output, vec![1, 2, 3, 4, 0]);
+
+        // Read next 3 bytes from cursor to volatile slice (amount read limited by length of remaining data in cursor)
+        assert_eq!(
+            cursor
+                .read_volatile(&mut VolatileSlice::from(&mut output[..4]))
+                .unwrap(),
+            3
+        );
+        assert_eq!(output, vec![5, 6, 7, 4, 0]);
+
+        cursor.set_position(0);
+        // Same as first test above, but with read_exact
+        cursor
+            .read_exact_volatile(&mut VolatileSlice::from(&mut output[..4]))
+            .unwrap();
+        assert_eq!(output, vec![1, 2, 3, 4, 0]);
+
+        // Same as above, but with read_exact. Should fail now, because we cannot fill a 4 byte buffer
+        // with whats remaining in the cursor (3 bytes). Output should remain unchanged.
+        assert!(cursor
+            .read_exact_volatile(&mut VolatileSlice::from(&mut output[..4]))
+            .is_err());
+        assert_eq!(output, vec![1, 2, 3, 4, 0]);
+    }
+
+    #[test]
+    fn test_write_volatile_for_cursor() {
+        let mut write_buffer = vec![0u8; 7];
+        let mut input = [1, 2, 3, 4];
+
+        let mut cursor = Cursor::new(write_buffer.as_mut_slice());
+
+        // Write 4 bytes from volatile slice to cursor (amount written limited by volatile slice length)
+        assert_eq!(
+            cursor
+                .write_volatile(&VolatileSlice::from(input.as_mut_slice()))
+                .unwrap(),
+            4
+        );
+        assert_eq!(cursor.get_ref(), &[1, 2, 3, 4, 0, 0, 0]);
+
+        // Write 3 bytes from volatile slice to cursor (amount written limited by remaining space in cursor)
+        assert_eq!(
+            cursor
+                .write_volatile(&VolatileSlice::from(input.as_mut_slice()))
+                .unwrap(),
+            3
+        );
+        assert_eq!(cursor.get_ref(), &[1, 2, 3, 4, 1, 2, 3]);
+    }
+
+    #[test]
+    fn test_write_volatile_for_vec() {
+        let mut write_buffer = Vec::new();
+        let mut input = [1, 2, 3, 4];
+
+        assert_eq!(
+            write_buffer
+                .write_volatile(&VolatileSlice::from(input.as_mut_slice()))
+                .unwrap(),
+            4
+        );
+
+        assert_eq!(&write_buffer, &input);
     }
 }